% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc1.R
\name{fnc}
\alias{fnc}
\title{Elementwise Square}
\usage{
fnc(x)
}
\arguments{
\item{x}{A numeric vector (or scalar).}
}
\value{
The output is a vector with the same length as \code{x}.
}
\description{
\code{fnc} returns square of each element of a numeric vector.
}
\examples{
fnc(c(1,2,4))
\dontrun{
  fnc("a")
}
}
