% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halve_cats.R
\name{halve_cats}
\alias{halve_cats}
\title{Cat Sharing}
\usage{
halve_cats(ncats)
}
\arguments{
\item{ncats}{number of cats to share}
}
\description{
Wrapper for the  'halve_cats' C function
}
\examples{
halve_cats(5)
}
\keyword{cats}

