% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try1.R
\name{MyCorr}
\alias{MyCorr}
\title{Estimate and visualize correlation between two numerical variables}
\usage{
MyCorr(x, y)
}
\arguments{
\item{x}{numerical vector of the first variable}

\item{y}{numerical vector of the second variable}
}
\description{
Estimate and visualize correlation between two numerical variables
}
